+++ src/ThunderEgg/GMG/LinearRestrictor.h
@@ -30,7 +30,7 @@ namespace GMG
 /**
  * @brief Restrictor that averages the corresponding fine cells into each coarse cell.
  */
-template <size_t D> class LinearRestrictor : public MPIRestrictor<D>
+template <int D> class LinearRestrictor : public MPIRestrictor<D>
 {
 	private:
 	void
