+++ src/ThunderEgg/GMG/VCycle.h
@@ -30,7 +30,7 @@ namespace GMG
 /**
  * @brief Implementation of a V-cycle
  */
-template <size_t D> class VCycle : public Cycle<D>
+template <int D> class VCycle : public Cycle<D>
 {
 	private:
 	int num_pre_sweeps    = 1;
@@ -60,7 +60,9 @@ template <size_t D> class VCycle : public Cycle<D>
 				this->smooth(level, u_vectors, f_vectors);
 			}
 		}
-		if (!level.finest()) { this->prepFiner(level, u_vectors, f_vectors); }
+		if (!level.finest()) {
+			this->prepFiner(level, u_vectors, f_vectors);
+		}
 	}
 
 	public:
